[server] Automatically set GOMAXPROCS to match CPU quota.
This is apparently considered a best practice when running
stuff on GKE (since by default GOMAXPROCS is set to the number
of cores on the host node, which results in a lot of contention
if the container is given small CPU quota).

Looks like it should work on Cloud Run as well.

See https://medium.com/@aditimishra_541/go-performance-with-ubers-automaxprocs-8f31226a92cd

R=gregorynisbet@google.com

Change-Id: Idd5a5aa949dd1f115f7c1df393d39f9c5faa625d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6448713
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>