[cv] ignore the default attention set rule when resetting CQ vote
When it resets the CQ vote, it removes the voter from the attention
set by the following default rule, because CV resets it on behalf
of the user.

> Replying (commenting, voting or just writing a change message)
> removes the replying user from the attention set

This CL is to ignore the default rule to keep the voter
in the attention set.

Bug: 409083087
Change-Id: I639b523aaf4be42f7a22d53c3f2ee828d265b432
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6444070
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>