[stacktag] Add a specific stacktrace error Tag.
This is built on debugstack to ensure that we don't see extra annoying
frames.
  * implements a merge function to pick the longest stack when confrontend with a multierror/joined error.
  * implements a special case so that sentinel errors will not have trivial/useless stacks on them.

If we decide that we need to see ALL the stacks, we can later change
Capture to just do a capture on every call, and maintain a map of
goroutine ID -> stack, just like we used to, but I want to see if the
simpler approach is sufficient for now.

R=fancl@chromium.org, gregorynisbet, vadimsh, vadimsh@chromium.org

Change-Id: Ic51244f26518334614b3a05c48c57a7cbba07e3a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6321230
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>