[reauth] Remove AppID check
It seems like there are at least two AppIDs in use.  Upon
re-evaluation, the risk of MITM is effectively zero since we SSL to a
static known Google URI for ReAuth, so checking the AppID is pointless
and may break if we don't cover the full set of valid values.

Bug: 390280795
Change-Id: I2d7965c2dfe34ff98b8dce8e2e9352cb5880ccf7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6484212
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Allen Li <ayatane@chromium.org>