[lucicfg] Terminal UI for displaying concurrent activities.
It is enabled only of stderr is a terminal and logging
level is Info. Otherwise it regresses to simple logging.

Another noteworthy feature is that it waits for 1s before
showing anything, to avoid excessive sensory overload in
(frequent) case the cache is populated and all operations
are super quick. It pops up only if lucicfg gets stuck,
to explain why it is stuck.

More polish will likely be needed (in particular around
error and progress reporting), but it is a start.

R=iannucci@chromium.org

Change-Id: Idb896e69cc326e933d25f4539b8b04272a03b173
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6444469
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@google.com>