[common/errors] Add generated source transformer tool.
Generated this tool from the prompt in the comment.

Tweaked it a bit to pull a list of filenames from stdin instead of
as a CLI flag, and cleaned up the code to remove a bit of extraneous
junk.

R=chanli, gregorynisbet@google.com, vadimsh

Change-Id: Ib016ed48bc4fedc5dbddeb5874472d605c6b5f35
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6486500
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>