Drop dependency on deprecated yaml
It is easier than I expected to replace everything with a call to
https://github.com/goccy/go-yaml .

This CL is *dangerous* though and we might want to break it up, depending
on the release schedule of the services affected.

In a few places we depend on the yes and no YAML booleans, and those
are not supported in go-yaml except when parsing something statically
known to be a bool. This can bite us if we aren't careful.

Change-Id: Ic12bc1b8c1c849019183b198274e65891822ec93
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6496009
Reviewed-by: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Robbie Iannucci <iannucci@google.com>