[ResultDB] Normalise failure reason storage around Errors collection.
The Errors collection should be the source of truth for the
failure reason's errors, with primary_error_message becoming
a computed field. Of the two, only the errors collection should be
stored in Spanner.

At upload time, we convert the primary_error_message into
a single-entry errors collection if errors is unset, to support
handling legacy clients.

This CL is best read in context of the validation CL, in partcular
the new ValidateFailureReason method:
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6450325

BUG=b:410660296

Change-Id: I3746e48f8708e9af76785cabd82f75d970b2b464
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6477726
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>