[errors] Add errors.WrapIf.
As we migrate away from errors.Annotate, it's clear there are a lot
of call sites which assume that the error wrapping function knows to
check `err` for nil and be a no-op if it is.

fmt.Errorf doesn't have this behavior - fmt.Errorf("%w", nil) always
returns a non-nil error.

The new `errors.WrapIf` function meets these two worlds in the
middle - still a very slim wrapper around the standard way of
annotating errors (with fmt.Errorf), but preserve the nil-checking
and "context: wrapped" functionality that errors.Annotate provided,
but without the weird intermediate structure, or the need to call
`Err()` at the end.

R=chanli, gregorynisbet@google.com, vadimsh

Change-Id: Iab928fd08a64d0fb15943342752a747e9ab22a41
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6496451
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Robbie Iannucci <iannucci@google.com>