[auth,config,grpc,led] Fix callsites where errors.Annotate is called with possible nil.
Fix callsites of errors.Annotate where the error argument is not
unambiguously non-nil.

An upcoming change will rewrite all:

  errors.Annotate(err, fmt, args...).Err()

to (effectively, but generally not with the explicit string addition):

  errors.Fmt(fmt+": %w", args..., err)

`errors.Fmt` is a thin wrapper around the stdlib `fmt.Errorf` which always
returns a non-nil error, even if the given `err` is nil. The wrapper
just ensures that a stack trace is attached to the returned error.

`errors.WrapIf` is similar to `errors.Fmt`, but returns nil if the
wrapped error is nil.

This is part of the cleanup effort to remove `errors.Annotate` and
`errors.Reason` to replace them with the more standard `errors.Fmt`
and `errors.New` (and `errors.WrapIf` if appropriate for the
replaced context).

R=chanli, gregorynisbet@google.com, vadimsh

Change-Id: Ia5c29ea66d80c6a7dc411b659d02bcd749c6cdcc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6496414
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>