[buildbucket] Fix callsites where errors.Annotate is called with possible nil.
Fix callsites of errors.Annotate where the error argument is not
unambiguously non-nil.

An upcoming change will rewrite all:

  errors.Annotate(err, fmt, args...).Err()

to (effectively, but generally not with the explicit string addition):

  errors.Fmt(fmt+": %w", args..., err)

`errors.Fmt` is a thin wrapper around the stdlib `fmt.Errorf` which always
returns a non-nil error, even if the given `err` is nil. The wrapper
just ensures that a stack trace is attached to the returned error.

`errors.WrapIf` is similar to `errors.Fmt`, but returns nil if the
wrapped error is nil.

This is part of the cleanup effort to remove `errors.Annotate` and
`errors.Reason` to replace them with the more standard `errors.Fmt`
and `errors.New` (and `errors.WrapIf` if appropriate for the
replaced context).

R=chanli, gregorynisbet@google.com, vadimsh

Change-Id: I388e2cf0045005031922fb739fb64b114947f0da
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6500066
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@google.com>