luci ui: fill out new test investigate page
This was coded very quickly, there are many things here that may be
inaccurate and definitely do not conform to the design.  It represents a
base to start iterating from, not a finished product.

That said, it now displays enough test info to be useful in debugging
test failures, and is probably ready for evaluation before starting
trusted tester phase.

Bug: 403013553
Change-Id: I2b046da0b79a381a5c114ecea7ddd6ea47ebb638
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6515134
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>