[auth] Add "RequiresWarmup" aspect of a token provider.
When set, it means we should try to get a token as soon as possible
(when constructing the client), instead of delaying the call until
the token is actually needed (e.g. during the first RPC).

This will be used for external credential provider, since calling
them is the only way to verify they are configured correctly (e.g.
the user has signed in).

BUG=b/414989137

Change-Id: If6fd06a8ca11d9d1099c7b41c5540f1a4f04b5b9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6514511
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@google.com>