[swarming] Add TaskResultSummary.SliceExpiration entity field.
It just identically set to TaskToRun.Expiration of the current
pending slice.

It will be used in "report-tasks" cron job to expire slices
that are stuck due to delayed or missing RBE PubSub notifications.

Note this is different from Python version, which runs a separate
cron job that scans TaskToRun entities directly using a datastore
index. This is an overkill for the Go version, because:
  1. We already have a cron job that scans all pending (and running)
     tasks. It scans TaskResultSummary entities though, which is
     why this CL is required.
  2. We expect vast majority of expirations happen through RBE
     PubSub notifications. This is just a fallback mechanism. Some
     latency (in particular latency of skipping running tasks in
     the scanning cron job) is acceptable.

This change would allows us to remove an index on TaskToRun.expiration
(and unshard the entity kind). This was once one of the hottest
indices in Swarming datastore. TaskToRun doesn't really need any
indices anymore (and thus sharding it makes no sense).

R=chanli@chromium.org

Change-Id: I83b464de83c2f1df797317fa2aadf2d1ef3b5b9c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6548333
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>