ui: Disable text-artifact that is crashing the browser
Something in the new version of react appears to disagree with the
test-artifact tag which is causing an infinite render loop, hanging the
browser.

Temporarily disable this until we can find a fix.

Workaround - open the text artifacts from the artifact list

this is preferred to rolling back the update patch which took many weeks
to get stable

Bug: 418030855
Change-Id: I0e435258e5c87167fd590b1e600eefdfd1ec395a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6552598
Auto-Submit: Matthew Warton <mwarton@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>