[ResultDB] Add order_by support to QueryTestVariants.
Allow the UI to sort by (effective) verdict status v2 instead
of the verdict status v1.

This implementation does come with a caveat, which is that the
status which we do not sort by may not always be correct (e.g.
verdict status v1 may not be correct if we sort by v2 status and
where the number of results in the verdict is more than the
ResultLimit). Also, PASSED and SKIPPED results have the same
sort priority for performance reasons.

BUG=b:415878114
TEST=Integration tests

Change-Id: I5eb923e78c95e120c80541a3a4a771b5b1e732ec
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6545445
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>