[Analysis] Update QueryStats to no longer require backfill of StatusV2.
The backfill of StatusV2 in the TestResults table is taking too long
to run and is blocking release of the MILO UI changes which depend
on the QueryStats API returning correct data. Introduce a work around
that allows us to proceed without the backfill.

BUG=b:415878114
TEST=Integration tests

Change-Id: Id6e43ceb1983be8d6ffe4ed9c2c24d1594440f2c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6577201
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>