[MILO] Update test history UI to new verdict statuses.
To preserve the behaviour of the old UI as best as possible in
the new status model, both PASSED and SKIPPED counts do not appear
by default and an icon is displayed in a cell instead (if all are
passed or all are skipped).

A new behaviour is that counts for passed, or skips, or both
can be enabled. This is useful if e.g. one wants to know the
total number of results on a day.

BUG=b:415878114
TEST=npm run test
TEST=spin up UI locally and smoke test

Change-Id: I137b36e7f52a5d7fb0d6151c70dd879f70e0a943
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6549658
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>