[MILO] Switch more endpoints to using status v2-based verdict order.
Due to limitations of the QueryTestVariants implementation, the
correct verdict status may not be returned for verdicts with many
test results if the sort order is not based on that status. As UIs
are now showing v2 status, use that sort order.

BUG=b:415878114
TEST=Integration tests

Change-Id: I09a0d11e9c662f13f51827cd0aaa4733677819a4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6585573
Auto-Submit: Patrick Meiring <meiring@google.com>
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Mostafa Draz <mdraz@google.com>