[ResultDB] Control test results order for QueryTestVariants.
Within a test variant, control the order of results so that the
order is:
1. FAILED
2. PASSED
3. SKIPPED
4. EXECUTION ERRORED
5. PRECLUDED

This approximately aligns with the strength of the results in
determining the test verdict. E.g. FAILED and PASSED results
are stronger than SKIPPED results for determining the verdict
status. Also, between passed and failed results, failed are more
interesting so we should prefer to show those.

The change in test expectations for reproduction instructions
owes to the fact that the implementation determines instructions
based on the first test result, and this CL changes what the
first result is.

BUG=b:409801263,b:415878114
TEST=Integration tests

Change-Id: I6b62ef48d6473e6937ec008217b8582a1b0da89a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6585574
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>