[errors] Fix possibly-nil err annotations in cv.
This is finding all locations where `errors.Annotate` is called with
an error that MIGHT be `nil`. The replacements are:
  * `errors.WrapIf` - the err specifically could be nil, and we just
    want to do the `if err != nil { <wrap> }` thing.
  * `errors.Fmt` - the err is NOT ever nil, and we want to
    unconditionally format it.

This is for an upcoming change to remove `errors.Annotate` entirely,
but `errors.Fmt` (the automatic replacement for `errors.Annotate`)
will unconditionally return a non-nil error, even if it's told to
wrap `nil`.

R=gregorynisbet, vadimsh, yiwzhang

Change-Id: Ia659ca30e33422b73c8529c8db3dd655efc0ae9d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6594765
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Robbie Iannucci <iannucci@google.com>