[errors] Fix possibly-nil err annotations in resultdb.
This is finding all locations where `errors.Annotate` is called with
an error that MIGHT be `nil`. The replacements are:
  * `errors.WrapIf` - the err specifically could be nil, and we just
    want to do the `if err != nil { <wrap> }` thing.
  * `errors.Fmt` - the err is NOT ever nil, and we want to
    unconditionally format it.

This is for an upcoming change to remove `errors.Annotate` entirely,
but `errors.Fmt` (the automatic replacement for `errors.Annotate`)
will unconditionally return a non-nil error, even if it's told to
wrap `nil`.

R=gregorynisbet, meiring, vadimsh

Change-Id: I673a01ea8df5f801e9eb722c2879bdcaa4263c24
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6595188
Auto-Submit: Robbie Iannucci <iannucci@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>