[cipd] Allow using "long" unix domain socket paths.
Unix domain socket path length is limited by 108 bytes on
Linux (and something similar on OSX).

Use symlinks to shorten the path to the directory that contains
the socket. Note that we can't symlink the socket itself, since
then the FS object is not a socket anymore (but a symlink).

On Linux try to use an existing /proc/self/fd/... symlink first
to avoid creating any new symlinks or "guessing" existence of "/tmp".

If this doesn't work, then create a symlink in hardcoded "/tmp". We
can't use $TMPDIR because it is generally too long (this is what
CIPD was doing previously and it was hitting path length limits).

This trick was borrowed from Soong:
https://android.googlesource.com/platform/build/soong/+/18490113/ui/build/paths/logs.go

BUG=b/419367985
R=iannucci@chromium.org

Change-Id: I8c45136d2700b4bd71d1189400a823d3dae8397d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6598975
Commit-Queue: Robbie Iannucci <iannucci@google.com>
Reviewed-by: Robbie Iannucci <iannucci@google.com>