[auth-service] Pull getting latest configs out of collateLatestRoles
This refactor will help with analyzing the latest configs for realms
data. There is no reason to fetch the latest configs twice - we can
just fetch them outside of the roles collation, and pass them again
when we analyze realms (in a following CL).

Bug: b/418865388
Change-Id: I5dc12430b2e72ec0bf337739842fa485127c1524
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6600357
Reviewed-by: Rachael Newitt <renewitt@google.com>
Commit-Queue: Anne Redulla <aredulla@google.com>