resultdb: Add AIP-160 filtering to QueryTestVariants
Only exposing three columns for now, test_id, variant and variant_hash.

The other obvious one is status, but this is more complicated due to v1,
v2 and overrides.  Just leaving the existing predicate to control status
for now.

This is to support filtering on the invocation page of the
test-investigate UI.

Bug: 422272789
Change-Id: Id65d9bd304d10b07d746503bbf8a1a00cd41ad71
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6694870
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>