ui: Add test id filtering to invocation page
Add test id filtering chip to the invocation page on the
test-investigate page.  This is  blocking feedback item for enabling a
wider distribution of the page.

Variant filtering is also available from the underlying backend, but the
UI is not implemented in this CL to limit the CL size.

Bug: 422272789
Change-Id: I67cb47f21c8e645427df732d9849ea51092d2fc3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6703689
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Heidi Chan <heidichan@google.com>