[cv] use empty identity if trigger.Email is empty
There happenes to be cases where CQ vote value is set but the CL
snapshot is missing the triggerer email. It causes a retry failure loop
so that the Run stays in pending, and a new CL event doesn't stop
the retry loop.

It actually makes sense. CV retries on most of internal errors, and
CQ votes without the triggerer email info is an unexpected error
condition. Hence, it keeps retrying.

A new CL change event would trigger the CL entity snapshot to be updated
and that actually happened in the particular CL where this error was
found. However, the logic evaluates the triggerer information from
RunCL entity, which is embedded with the CL snapshot evaluated at
the time of the Run creation. It's intended to be frozen, and it makes
sense the updated CL snapshot should just result in the Run to be
cancelled by a separate pipeline, but somehow it's not happending, when
the triggerer is missed and the Run couldn't event start. Fixing it would
be a separate CL.

This CL patches the check such that it'd just leave the triggerer
identity empty so that the CQ run would be rejected by isDryunner() or
isCommitter() w/o the retry loop.

Bug: b/428754990
Change-Id: If51b15268bf9c27c3c60ede80bb03314281ba429
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6730956
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Gavin Mak <gavinmak@google.com>