[git-credential-luci] Don't block reauth on enable
Since this is a separate command, we don't need to block it on the
envvar.  We do however print a notice in case the user wanted to use
reauth in get.  (The envvar will also be documented in setup guides
and the auto setup tool in depot_tools.)

Bug: 410936312
Change-Id: I4c113a7a2fd8a93ce954f66aaba724e8f7b9509e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6761974
Reviewed-by: Joey Scarr <jsca@google.com>
Commit-Queue: Allen Li <ayatane@chromium.org>