[Bisection] Only analyses failures with pass in the last 7 days
We are seeing a lot of INSUFFICIENTDATA in bisection UI https://ci.chromium.org/ui/p/chromium/bisection/test-analysis

This was because the last pass was not in the past 30 days, so we cannot get the build information.

This is WAI, but polluted the bisection UI.

This CL change the behaviour.

Old behavior
"We only consider failures that has a failure in the past 24 hours and had started failing in the past 30 days"

New behavior
"We only consider failures that has a failure in the past 24 hours and had the last pass in the past 7 days"



Change-Id: I3413477de846921b0441ff35d2b54f5f323fda39
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6768232
Auto-Submit: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Beining Chen <beining@google.com>