cv: mark a CL as unwatched if no ConfigGroups are applicable
evalcl.go:makePCL() leaves PCL.ConfigGroupIndexes empty if all of below
conditions are met
- the CL was previously watched so that it has ApplicableConfigGroup
  in the DS entity.
- the ApplicableConfigGroup was removed, and none of other ConfigGroups
  in the project config are applicable for the CL.

In the case, makePCL() sets the status with OK while
leaving ConfigGroupIndexes empty. It's a bug. It should have set
the status with UNWATCHED.

However, the incorrect PCL status doesn't cause issues in most cases,
because the triageCL checks the length of it and considers it as
UNWATCHED, if empty.
- https://source.chromium.org/chromium/infra/infra_superproject/+/main:infra/go/src/go.chromium.org/luci/cv/internal/prjmanager/triager/cls.go;l=200

However, it becomes an issue when PM triages dependent CLs
in a combined mode. (aka chromeos mode)
In the mode, dependent CLs are not triaged as an individual CL, but
triaged together via TriagedDeps:categorize(). It assumes that
all dep CLs have ConfigGroupIndexes and panics otherwise.
- https://source.chromium.org/chromium/infra/infra_superproject/+/main:infra/go/src/go.chromium.org/luci/cv/internal/prjmanager/triager/deps.go;l=142;drc=74445425c4254e61773f3576f28f2f93b2b2c31b

This CL updates the evaluation logic of PCL so that the PCL entity
for the CL will be marked as UNWATCHED if the dep CL is no longer
watched by any CQ group. Then, the dep would be considered as
an invalid dep and the CQ vote would be revoked with a message
indicating the reason for the revoke.
- https://source.chromium.org/chromium/infra/infra_superproject/+/main:infra/go/src/go.chromium.org/luci/cv/internal/prjmanager/triager/deps.go;l=117-122;drc=74445425c4254e61773f3576f28f2f93b2b2c31b

Bug: 40196810
Change-Id: I402948f1fda8d63490a7ae56bac55341cd4d8120
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6771210
Commit-Queue: Brian Ryner <bryner@google.com>
Auto-Submit: Scott Lee <ddoman@chromium.org>
Reviewed-by: Brian Ryner <bryner@google.com>