Accept list of GridColDef in useColumnManagement instead of building it to account for differences between various tables
This commit fixes a regression I introduced when implementing the `useColumnManagement` hook and applying it to the `ResourceRequestInsightsPage`.

My initial implementation had the hook generate its own columns by calling the `getColumns` utility. I had incorrectly assumed this utility would be generic enough to work for both the device table and
the resource request table. This was an oversight, as `getColumns` is tightly coupled to the device data structure and cannot handle the schema for resource requests. This caused a rendering crash on
the insights page.

The fix refactors `useColumnManagement` to be more data-agnostic. It no longer generates columns internally and instead accepts a pre-configured array of `GridColDef` objects as a parameter.
Both the `DeviceTable` and `ResourceRequestInsightsPage` are updated to provide their own specific column configurations to the hook.

Change-Id: I302dafc77e441da2f8ad5e22c42676597ff2bfbe
Bug:433968040
Demo URL: https://18677-d32902b-tidianed-fix-columns-management-dot-luci-milo-dev.appspot.com/ui/fleet/labs/requests
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6785368
Reviewed-by: Piotr Grabski-Gradziński <piotrgg@google.com>
Commit-Queue: Tidiane Dia <tidianedia@google.com>