[ResultDB] Refine FinalizeRootInvocationRequest.
Require callers to specify whether to finalize the root work unit
with the root invocation. Callers must specify the desired behaviour
explicitly to ensure they are aware of the need to call (or not call)
FinalizeWorkUnit for the root work unit.

Callers can choose between matching Create and Finalize calls 1 for 1,
or matching a Finalize call to each created resource 1 for 1.

BUG=b:422275479
TEST=None

Change-Id: Ia2978120d1530c278c6c6cfa5c49247bbae690b6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6796348
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>