[ResultDB] Refactor WorkUnit.Read, add child- work unit and inv index.
Reduce three read calls by the business logic down to one. Also
reduce three Spanner queries down into one, which will touch only
~one Spanner shard.

The first change is for simplicity and to reduce the chance of bugs;
the second change is to take out some scale insurance as the
current method to listing child work units could be expensive if the
number of work units in an invocation is large.

This comes at the cost of writing data to an extra place when we
create work units or work unit inclusions.

BUG=None
TEST=Integration tests

Change-Id: Ic4027e3b9fef256538a9944561ed4fa7c22fb48b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6790587
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>