[ResultDB] Add module_id to work units and invocations.
This is necessary for associating work unit-level errors with
modules (for test result aggregations). The requirement for the
module to be set before upload helps ensure any test execution
errors will be bubbled up to the correct modules.

It also facilitates validation with the uploaded test results.

The field is also added to the invocations table to allow enforcing
constraints early and so de-risk the future invocations ->
work units migration.

BUG=b:434107587
TEST=None

Change-Id: I64324ae6bf889cd747e00a7119071f01bdaad073
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6801106
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>