LUCI UI: Improve error reporting by unminifying error stacks
Currently we use stackdriver-errors-js lib to send errors to the error
reporting API. This lib uses couple other libs that had some issues
which I described in the related bug. Also it wasn't properly handling
relative paths which was one of the causes that it couldn't use the
sourcemaps to unminify the call stacks.

In this CL I address that issues.

Bug: b/431711855
Change-Id: Ie35a0ddd804c4edf5c51e138852c37ced9978aae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6797851
Reviewed-by: Konrad Polarczyk <polarczyk@google.com>
Commit-Queue: Vaghinak Vardanyan <vaghinak@google.com>