LUCI UI: Improve reported error stack frames
source-map lib unminifies the stack frame and resolves the original
position really good. But it struggles with the enclosing function
names. Actually it is not that easy to resolve given just the
surrounding lines of the code. Here I manually find the enclosing
function. Tested it manually, feels it is good enough. We can improve it
gradually in the future to cover even more paths if needed.

Bug: b/431711855
Change-Id: I0adfcb2a6b3bacdb03190acfcdb7d315a1ec704c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6797852
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Vaghinak Vardanyan <vaghinak@google.com>