fleet console ui: move some of the task table helper functions into a utils file
This is a preparation step so the helper functions in device details can
be used in for the admin tasks history page.

Testing: manually ran the UI server -
https://screenshot.googleplex.com/4286tBbHTfZhDAy . Also, helper
functions are not exported in original file so other components outside
of tasks_table should not depend on it.

Bug: b/431372703

Change-Id: Iec6ece5565f8764f97abd931b76f403176a14834
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6807035
Reviewed-by: Bryan Ni <bniche@google.com>
Commit-Queue: Phoebe Tang <phoebetang@google.com>