[ResultDB] Set module_id in ResultSink.
This is nominally a breaking change, as it constrains the
invocation to only upload one module whereas previously
any number was possible. However, there are only a few sites
which do this, which are in infra project, and we will update
the recipes to pass -new.

BUG=b:434107587

Change-Id: I56bd43b9d91143eba73947a1df2490f8d2cabb37
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6810665
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>