LUCI UI: Change imports back to the placeholders
importmap helps to prevent cascading cache invalidation but we need to
use the placeholder in the imports instead of the actual file names. In
the previous CL that fixed the sourcemaps issue in the generateBundle
stage I missed that part and was using actual file names so even though
the importmap was created it wasn't being used. This Cl addresses that
issue.

Bug: b/431711855
Change-Id: I06dd4a56c9491e7831828ab4070a0f7c7ce7f2bb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6820260
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Vaghinak Vardanyan <vaghinak@google.com>