[ResultDB] Add LastUpdated time to WorkUnits and RootInvocations.
This will facilitate adding etags to these records in future
for resource freshness validation prior to updates and for caching.

The scope of LastUpdated time is anything that might appear on the
RootInvocation or WorkUnit proto; for work units, this means a change
to child work units and invocations will also need to trigger an update
to the time.

BUG=b:434828469
TEST=Integration tests

Change-Id: Ieed5f5c18232e091f64eb69e65aaf1c72bd174b1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6821104
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>