[ResultDB] Update BatchCreateArtifacts to accept variant.
Simplify and standardise the API surface by using the same
TestIdentifier message on test results and artifacts.

This should be API compatible as the field names and numbers
match between the TestIdentifier message and the deleted
TestIdentifierBase message.

Long-run, the variant stored on the artifact will always match
that on the work unit, but in Spanner it may still be stored
denormalised to support queries (e.g. within a root invocation).

BUG=b:434107587
TEST=Integration tests

Change-Id: I61fd6ee1166f13f14c935b70d0b1330437548c99
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6821103
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>