LUCI UI: Attempt to make import_map_plugin.ts more stable by sorting chunks
There is a chance there the order of the chunks in the bundle could be affecting how our builds are being processed across different environments because the order is not guaranteed.

This change adds a stable order for our chunk processing in our import map logic to eliminate this as a possible source of differences between environments.

Bug:b/436152777
Change-Id: I33b97ef89934b5ed827476705c98447470d68396
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6825763
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>