Invalidate tasks query after Autorepair dialog closes.
When the Autorepair dialog is closed after a session has results, invalidate the 'tasks' query to refresh the task list. This ensures that the newly created repair tasks are visible to the user. Also, reset the session information and deep repair state when the dialog is closed.

Change-Id: I6bacb38c16115986339b8b9d9aacdd4af0c584c5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6847246
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>