[BHN] Make getBuilder calls async
Make calls async so that emails with larger number of builders
are run in a promptly manner. Also, this CL introduces a wait
to allow completion of all the getBuilder calls before moving forward.
This prevents any calls from being dropped or unfinished when we
are generating the email to send out to users.

Bug: 407539622
Change-Id: I3311940169a981b77c8135b03cef017fb2472b94
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6859159
Commit-Queue: Riley Wong <rgw@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>