monitoring: don't mislead the user with incorrect links
Since we only load a limited number of builds, if they are all failures
we cannot tell if the first failure is actually the first failure, or we
are just seeing part of a longer sequence of failures.

Also remove silence funcitonality that was not working.

Bug: 439689543
Change-Id: Ice68bf4de2b1cabe68292b67bbb4e8b953e82962
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6888133
Commit-Queue: Mostafa Draz <mdraz@google.com>
Auto-Submit: Matthew Warton <mwarton@google.com>
Reviewed-by: Mostafa Draz <mdraz@google.com>