[swarming] Implement 2 legacy API endpoints on top of new API.
Two endpoints:
  /_ah/api/swarming/v1/bots/count
  /_ah/api/swarming/v1/bots/list

It has been years since they were deprecated and callers were asked
to stop using them, but they are still in use on chromeos-swarming.
Just implement them on top of the new API to avoid delaying removal
of the Python portion of Swarming for several more years.

Do it YOLO-style without any traffic splitting, since setting it up
is non-trivial and whoever uses these endpoints should not be using
them to begin with, so if they break, it is a good signal to stop
using legacy APIs.

R=chanli@chromium.org

Change-Id: I39b060f092527dbb57ec36398ac1830e78659842
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6893057
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>