[ResultDB] Add Get methods to recorder interface.
The Get methods on the recorder side would be authorized with
the update token instead of the permission model, meaning
uploaders no longer need to be granted read access to the realm
they are uploading to but can operate with access to their work unit
only. This is not a privilege escalation however as Create, Update and Finalize RPCs already return the resource.

This has two motivations:
- Client ergonomics: recorder clients do not also need to create
  an instance of the read-side client.
- Separating failure domains: an overload or failure of the
  ResultDB-side service is less likely to cascade to failure of
  recorder clients as the traffic is handled by different backends.
  Of course this will only work for some uploaders; this will not
  help some recipes which are dependent on QueryTestVariants (e.g.
  to query what to exonerate).

BUG=None
TEST=Integration tests

Change-Id: I31efc7e073f8d406c034cef91b7badacd47c3839
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6898496
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>