[ResultDB] Add support for work units in ListArtifacts.
The goal of this RPC is to support functionality equivalent to the
current ListArtifacts for invocations, for work units.

Listing of masked artifacts is not supported, but listing artifacts
where the caller has artifacts.listLimited on the root invocation
and artifacts.get permission on the work unit is supported.
This is to get work unit granularity access controls, which
are a requirement for some existing users.

In a follow-up, we will update the artifact download
endpoint to support work units.

BUG=b:441369705
TEST=Integration tests

Change-Id: I6e64ad326ad782235182c09854f8ce37aac9d40e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6906703
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>