[common] Increase mathrand fairness bounds.
Test is flaky. It seems the test regressed at a prior point or
the seed became actually random - either way, the test is too flaky
and is blocking CQ runs.

It is not clear that perfect fairness is super important for the
insecure RNG as it is not used in any security-sensitive processes.

BUG=b:44268677
8TEST=Integration tests

Change-Id: I323c224116afef4a99bb44775f7eaf263fed8827
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6918031
Auto-Submit: Patrick Meiring <meiring@google.com>
Reviewed-by: Fumitoshi Ukai <ukai@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>